# Conflicts:
#	dune/xt/grid/exceptions.hh
#	dune/xt/grid/test/builder_definitions.cmake
#	dune/xt/grid/test/compiles_totals.pickle
#	dune/xt/grid/test/walker.cc
#	dune/xt/grid/walker.hh
#	dune/xt/grid/walker/apply-on.hh

+++ dune/xt/grid/dd/subdomains/factory.hh
@@ -7,6 +7,7 @@
 // Authors:
 //   Felix Schindler (2017)
 //   Rene Milk       (2017 - 2018)
+//   Tobias Leibner  (2018)
 
 #ifndef DUNE_XT_GRID_DD_SUBDOMAINS_FACTORY_HH
 #define DUNE_XT_GRID_DD_SUBDOMAINS_FACTORY_HH
@@ -245,7 +246,7 @@ private:
                             CodimSizesType& localCodimSizes)
     {
       // loop over all codim c subentities of the entity
-      for (size_t i = 0; i < entity.subEntities(c); ++i) {
+      for (unsigned int i = 0; i < entity.subEntities(c); ++i) {
         const auto codimCentity = entity.template subEntity<c>(i);
         const GeometryType& geometryType = codimCentity.type();
         const IndexType globalIndex = factory.globalGridView_->indexSet().index(codimCentity);
@@ -908,7 +909,7 @@ struct SubdomainGridFactory<GridType>::Add<c, c>
                           typename SubdomainGridFactory<GridType>::CodimSizesType& localCodimSizes)
   {
     // loop over all codim c subentities of this entity
-    for (size_t i = 0; i < entity.subEntities(c); ++i) {
+    for (unsigned int i = 0; i < entity.subEntities(c); ++i) {
       const auto codimCentity = entity.template subEntity<c>(i);
       const SubdomainGridFactory<GridType>::GeometryType& geometryType = codimCentity.type();
       const typename SubdomainGridFactory<GridType>::IndexType globalIndex =
